test/requirements.txt
pytest>=7.0.0
pytest-html>=3.1.1
PyYAML>=6.0
# MySQL
peewee>=3.14.5
pymysql>=1.0.2